Initial support for drawings
this commits adds a lot of things needed to draw strokes into the
buffer.

It implements a "booktool" class for tools (i.e. pen, eraser etc.) to
inherit from. For now there is only a "Text" and "Pen" tool implemented,
but others will follow (eraser, select, maybe move).

The View now contains a "Stroke Proxy" that carries the state of the
active stroke, that is appended upon by the pen tool. This may move into
the pen tool class implementation, but not nesecressaraly, as other
tools (maybe caligraphic stroke) might be able to use it too.

Implements a MdNotebook.BoundDrawing Widget class and that contains a
bunch of MdNotebook.Strokes. Both have utiliy methods to make better
use out of them.

Implements a MdNotebook.Toolbar Widget that is able to let the View
switch between different BookTools and pen colors.